Fix vint4.set_lane on SSE2

+++ Source/astcenc_vecmathlib_sse_4.h
@@ -260,7 +260,7 @@ struct vint4
 #if ASTCENC_SSE >= 41
 		m = _mm_insert_epi32(m, a, l);
 #else
-		alignas(16) float idx[4];
+		alignas(16) int idx[4];
 		_mm_store_si128((__m128i*)idx, m);
 		idx[l] = a;
 		m = _mm_load_si128((const __m128i*)idx);
